+++ CoreAlgorithmPlugins/src/au/gov/asd/tac/constellation/plugins/algorithms/clustering/hierarchical/FastNewman.java
@@ -31,8 +31,10 @@ import java.util.Set;
 import java.util.TreeSet;
 
 /**
- * Execute the Fast Newman function which clusters the graph hierarchically by initially placing all vertices in their
- * own cluster and then iteratively merging clusters according to a weight function until the optimal state is reached.
+ * Execute the Fast Newman function which clusters the graph hierarchically by
+ * initially placing all vertices in their own cluster and then iteratively
+ * merging clusters according to a weight function until the optimal state is
+ * reached.
  *
  * @author sirius
  */
@@ -58,7 +60,7 @@ public class FastNewman {
         if (weightAttribute != null) {
             weightAttributeId = graph.getAttribute(GraphElementType.TRANSACTION, weightAttribute);
         }
-        double totalWeight = 0;
+        float totalWeight = 0;
         if (weightAttributeId != Graph.NOT_FOUND) {
             for (int p = 0; p < transactionCount; p++) {
                 final int transaction = graph.getTransaction(p);
@@ -78,7 +80,7 @@ public class FastNewman {
             groups[position] = new Group();
             groups[position].vertex = vxId;
             if (weightAttributeId == Graph.NOT_FOUND) {
-                groups[position].weight = (double) graph.getVertexTransactionCount(vxId) / transactionCount;
+                groups[position].weight = (float) graph.getVertexTransactionCount(vxId) / transactionCount;
             } else {
                 final int vertexTransactionCount = graph.getVertexTransactionCount(vxId);
                 for (int p = 0; p < vertexTransactionCount; p++) {
@@ -104,7 +106,7 @@ public class FastNewman {
                 link.lowGroup.links.put(link.highGroup, link);
 
                 if (weightAttributeId == Graph.NOT_FOUND) {
-                    link.weight = (double) graph.getLinkTransactionCount(linkId) / transactionCount;
+                    link.weight = (float) graph.getLinkTransactionCount(linkId) / transactionCount;
                 } else {
                     final int linkTransactionCount = graph.getLinkTransactionCount(linkId);
                     for (int tp = 0; tp < linkTransactionCount; tp++) {
@@ -113,7 +115,7 @@ public class FastNewman {
                     }
                     link.weight /= totalWeight;
                 }
-                link.deltaQ = 2d * (link.weight - link.highGroup.weight * link.lowGroup.weight);
+                link.deltaQ = 2f * (link.weight - link.highGroup.weight * link.lowGroup.weight);
 
                 link.initial = initialLinkIds.contains(linkId);
                 links.add(link);
@@ -244,7 +246,7 @@ public class FastNewman {
     public static class Group {
 
         private int vertex;
-        private double weight = 0.0;
+        private float weight = 0.0f;
         private Group parent = null;
         private int mergeStep = Integer.MAX_VALUE;
         private Map<Group, Link> links = new HashMap<>();
@@ -296,8 +298,8 @@ public class FastNewman {
 
         private Group highGroup;
         private Group lowGroup;
-        private double weight = 0.0f;
-        private double deltaQ;
+        private float weight = 0.0f;
+        private float deltaQ;
         private boolean initial = false;
 
         @Override
